Move snake-camelCase converter to lib
